Move RequestConfigurationFactory
